// src/app/enums/v-card-type3-email.enum.ts
enum VCardType3EmailEnum {
  INTERNET = 'INTERNET', // to indicate an Internet addressing type
  X400 = 'X400', // to indicate an X.400 addressing type
  PREF = 'PREF' // to indicate a preferred-use email address when more than one
                // is specified.
  // A non-standard value can also be specified.
}

export default VCardType3EmailEnum